requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argilla==1.8.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.26.76
botocore==1.29.76
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
deeplake==3.5.4
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
effdet==0.4.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
humbug==0.3.1
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.188
layoutparser==0.3.4
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathos==0.3.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==2.2.1
portalocker==2.7.0
pox==0.3.2
ppft==1.7.6.6
protobuf==4.23.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
pypandoc==1.11
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
s3transfer==0.6.1
safetensors==0.3.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
unstructured==0.7.1
unstructured-inference==0.5.1
urllib3==1.26.16
Wand==0.6.11
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
